Move organizer ui controls to it's own file
