+++ src/ClassicUO.Client/Game/UI/Gumps/AssistantGump.cs
@@ -6,6 +6,7 @@ using ClassicUO.Configuration;
 using ClassicUO.Game.Data;
 using ClassicUO.Game.Managers;
 using ClassicUO.Game.UI.Controls;
+using ClassicUO.Game.UI.Controls.AssistantGumpControls;
 using ClassicUO.Resources;
 using ClassicUO.Utility;
 using Microsoft.Xna.Framework;
@@ -827,300 +828,10 @@ public class AssistantGump : BaseOptionsGump
         MainContent.AddToLeft(CategoryButton("Organizer", page, MainContent.LeftWidth));
         MainContent.ResetRightSide();
 
-        // Left side content - organizer list
-        var leftSideContent = new LeftSideMenuRightSideContent(MainContent.RightWidth, MainContent.Height, (int)(MainContent.RightWidth * 0.28));
-        MainContent.AddToRight(leftSideContent, true, page);
-
-        // Add new organizer button
-        ModernButton addButton = new(0, 0, leftSideContent.LeftWidth, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Add Organizer", ThemeSettings.BUTTON_FONT_COLOR);
-        addButton.MouseUp += (sender, e) =>
-        {
-            var newConfig = OrganizerAgent.Instance?.NewOrganizerConfig();
-            if (newConfig != null)
-            {
-                var configButton = CreateOrganizerConfigButton(newConfig, leftSideContent);
-                leftSideContent.AddToLeft(configButton);
-                // Auto-select the new config
-                SelectOrganizerConfig(newConfig, leftSideContent, configButton);
-            }
-        };
-        leftSideContent.AddToLeft(addButton);
-
-        // Organizer configurations list
-        OrganizerConfig firstConfig = null;
-        if (OrganizerAgent.Instance?.OrganizerConfigs != null)
-        {
-            for (int i = 0; i < OrganizerAgent.Instance.OrganizerConfigs.Count; i++)
-            {
-                var config = OrganizerAgent.Instance.OrganizerConfigs[i];
-                if (firstConfig == null) firstConfig = config;
-
-                var configButton = CreateOrganizerConfigButton(config, leftSideContent);
-                leftSideContent.AddToLeft(configButton);
-            }
-        }
+        var organizerControl = new OrganizerControl(MainContent.RightWidth, MainContent.Height);
+        MainContent.AddToRight(organizerControl, true, page);
     }
 
-    private ModernButton CreateOrganizerConfigButton(OrganizerConfig config, LeftSideMenuRightSideContent leftSideContent)
-    {
-        var button = new ModernButton(0, 0, leftSideContent.LeftWidth, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, config.Name, ThemeSettings.BUTTON_FONT_COLOR);
-        button.MouseUp += (sender, e) =>
-        {
-            SelectOrganizerConfig(config, leftSideContent, button);
-        };
-        return button;
-    }
-
-    private void SelectOrganizerConfig(OrganizerConfig config, LeftSideMenuRightSideContent leftSideContent, ModernButton button)
-    {
-        // Clear right side
-        leftSideContent.RightArea.Clear();
-        leftSideContent.ResetRightSide();
-
-        // Add configuration details to right side
-        BuildOrganizerConfigDetails(config, leftSideContent, button);
-    }
-
-    private void BuildOrganizerConfigDetails(OrganizerConfig config, LeftSideMenuRightSideContent leftSideContent, ModernButton button)
-    {
-        HBoxContainer box = new HBoxContainer(50);
-        // Name input
-        InputFieldWithLabel nameInput = null;
-        nameInput = new InputFieldWithLabel("Name", ThemeSettings.INPUT_WIDTH, config.Name, false, (s, e) =>
-        {
-            config.Name = nameInput.Text;
-            button.TextLabel.SetText(nameInput.Text);
-        });
-        leftSideContent.AddToRight(nameInput);
-
-        // Enabled checkbox
-        var enabledCheckbox = new CheckboxWithLabel("Enabled", 0, config.Enabled, b => config.Enabled = b)
-        {
-            X = nameInput.X + nameInput.Width + 10,
-            Y = nameInput.Y
-        };
-        leftSideContent.AddToRight(enabledCheckbox, false);
-
-        // Run organizer button
-        ModernButton runButton = new(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "Run Organizer", Color.Lime)
-        {
-            IsSelected = true,
-            X = enabledCheckbox.X,
-            Y = enabledCheckbox.Y + 30
-        };
-        runButton.MouseUp += (sender, e) =>
-        {
-            OrganizerAgent.Instance?.RunOrganizer(config.Name);
-        };
-        leftSideContent.AddToRight(runButton);
-
-        //Dupe button
-        ModernButton dupeButton = new(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "Duplicate", Color.Cyan)
-        {
-            IsSelected = true,
-            X = runButton.X
-        };
-        dupeButton.Y = enabledCheckbox.Y + 20;
-        dupeButton.MouseUp += (sender, e) =>
-        {
-            var dupedConfig = OrganizerAgent.Instance?.DupeConfig(config);
-
-            if (dupedConfig != null)
-            {
-                var configButton = CreateOrganizerConfigButton(dupedConfig, leftSideContent);
-                leftSideContent.AddToLeft(configButton);
-            }
-        };
-        leftSideContent.AddToRight(dupeButton);
-
-        //Macro button
-        ModernButton MacroButton = new(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "+ Macro Button", Color.Yellow)
-        {
-            IsSelected = true,
-        };
-        MacroButton.MouseUp += (sender, e) =>
-        {
-            OrganizerAgent.Instance?.CreateOrganizerMacroButton(config.Name);
-            GameActions.Print($"Created Organizer Macro: Organizer: {config.Name}");
-        };
-        leftSideContent.AddToRight(MacroButton);
-
-        // Delete organizer button
-        ModernButton deleteButton = new(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "Delete Organizer", Color.Red)
-        {
-            IsSelected = true,
-            X = dupeButton.X + dupeButton.Width + 60,
-            Y = runButton.Y
-        };
-
-        deleteButton.MouseUp += (sender, e) =>
-        {
-            OrganizerAgent.Instance?.DeleteConfig(config);
-            button.Dispose();
-        };
-        leftSideContent.AddToRight(deleteButton, false);
-        // Target container button
-        ModernButton SourceButton = new(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "Set Source", ThemeSettings.BUTTON_FONT_COLOR)
-        {
-            IsSelected = true
-        };
-        SourceButton.MouseUp += (s, e) =>
-        {
-            GameActions.Print("Select [SOURCE] Container", 82);
-            TargetManager.LastTargetInfo.Clear();
-            var source = TargetHelper.TargetObject((source) =>
-            {
-                if (source == null || !SerialHelper.IsItem(source))
-                {
-                    GameActions.Print("Only items can be selected!");
-                    return;
-                }
-                config.SourceContSerial = source.Serial;
-                GameActions.Print($"Source container set to {source.Serial:X}", 63);
-                SelectOrganizerConfig(config, leftSideContent, button);
-            });
-        };
-        leftSideContent.AddToRight(SourceButton);
-
-        ModernButton DestButton = new(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "Set Destination", ThemeSettings.BUTTON_FONT_COLOR)
-        {
-            IsSelected = true,
-            X = SourceButton.X + SourceButton.Width + 60,
-            Y = SourceButton.Y
-        };
-        DestButton.MouseUp += (s, e) =>
-        {
-            GameActions.Print("Select [DESTINATION] Container", 82);
-            TargetManager.LastTargetInfo.Clear();
-            var destination = TargetHelper.TargetObject((destination) =>
-            {
-                if (destination == null || !SerialHelper.IsItem(destination))
-                {
-                    GameActions.Print("Only items can be selected!");
-                    return;
-                }
-                config.DestContSerial = destination.Serial;
-                GameActions.Print($"Destination container set to {destination.Serial:X}", 63);
-                SelectOrganizerConfig(config, leftSideContent, button);
-            });
-        };
-        leftSideContent.AddToRight(DestButton,false);
-
-
-        box.BlankLine();
-        // Current source container display
-        if (config.SourceContSerial != 0)
-        {
-            box.Add(TextBox.GetOne($"Source ({World.Items.Get(config.SourceContSerial)?.Name}, {config.SourceContSerial:X})", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE - 1, Color.White, TextBox.RTLOptions.Default()));
-        }
-        else
-        {
-            box.Add(TextBox.GetOne($"Source (Your backpack)", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE - 1, Color.DarkGreen, TextBox.RTLOptions.Default()));
-        }
-        box.BlankLine();
-        // Current destination container display
-        if (config.DestContSerial != 0)
-        {
-            box.Add(TextBox.GetOne($"Dest. ({World.Items.Get(config.DestContSerial)?.Name}, {config.DestContSerial:X})", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE - 1, Color.White, TextBox.RTLOptions.Default()));
-        }
-        leftSideContent.AddToRight(box);
-        leftSideContent.BlankLine();
-
-        // Items to organize label
-        Control c;
-        leftSideContent.AddToRight(c = TextBox.GetOne("Items to organize:                              Amount:", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, Color.White, TextBox.RTLOptions.Default()));
-        // Add item button with targeting
-        ModernButton addItemButton = new(0, 0, 120, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "Add Item", ThemeSettings.BUTTON_FONT_COLOR)
-        {
-            X = leftSideContent.RightWidth - 145,
-            Y = c.Y - 13
-        };
-        addItemButton.MouseUp += (sender, e) =>
-        {
-            TargetHelper.TargetObject((obj) =>
-            {
-                if (!SerialHelper.IsItem(obj))
-                {
-                    GameActions.Print("Only items can be added!");
-                    return;
-                }
-
-                var newItemConfig = config.NewItemConfig();
-                newItemConfig.Graphic = obj.Graphic;
-                newItemConfig.Hue = obj.Hue;
-
-                GameActions.Print($"Added item: Graphic {obj.Graphic:X}, Hue {obj.Hue:X}");
-
-                // Refresh the right side to show the new item
-                SelectOrganizerConfig(config, leftSideContent, button);
-            });
-        };
-        leftSideContent.AddToRight(addItemButton, false);
-
-        // Item configurations list
-        foreach (var itemConfig in config.ItemConfigs)
-        {
-            var itemArea = CreateOrganizerItemConfigArea(itemConfig, config, leftSideContent, button);
-            leftSideContent.AddToRight(itemArea);
-        }
-    }
-
-    private Control CreateOrganizerItemConfigArea(OrganizerItemConfig itemConfig, OrganizerConfig parentConfig, LeftSideMenuRightSideContent leftSideContent, ModernButton button)
-    {
-        var itemArea = new Area() { Width = leftSideContent.RightArea.Width - 20, AcceptMouseInput = false };
-        Control c;
-
-        var rsp = new ResizableStaticPic(itemConfig.Graphic, 50, 50) { Hue = (ushort)(itemConfig.Hue == ushort.MaxValue ? 0 : itemConfig.Hue) };
-        itemArea.Add(rsp);
-
-        // Item info display
-        var itemText = $"Graphic: {itemConfig.Graphic:X4}, Hue: {(itemConfig.Hue == ushort.MaxValue ? "ANY" : itemConfig.Hue.ToString())}";
-        itemArea.Add(c = TextBox.GetOne(itemText, ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, Color.White, TextBox.RTLOptions.Default()));
-        c.X = rsp.Width + 1;
-        c.Y = (rsp.Height - c.Height) / 2;
-
-        InputField input = new InputField(100, ThemeSettings.CHECKBOX_SIZE, 100, 60000, itemConfig.Amount.ToString(), true,
-        onTextChanges: (s, e) =>
-        {
-            if (ushort.TryParse(((InputField.StbTextBox)s).Text, out ushort amount))
-            {
-                if (amount <= 0) amount = 0;
-                itemConfig.Amount = amount;
-            }
-        })
-        {
-            X = 240,
-            Y = 10
-        };
-
-        input.SetTooltip("Amount of item quantity to move, 0 = All");
-
-        itemArea.Add(input);
-        // Enabled checkbox
-        var enabledCheckbox = new CheckboxWithLabel("Enabled", 0, itemConfig.Enabled, b => itemConfig.Enabled = b)
-        {
-            X = itemArea.Width - 130
-        };
-        enabledCheckbox.Y = (rsp.Height - enabledCheckbox.Height) / 2;
-        itemArea.Add(enabledCheckbox);
-
-        // Delete button
-        ModernButton deleteButton = new(0, 0, 20, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "X", Color.Red)
-        {
-            X = enabledCheckbox.X + enabledCheckbox.Width + 10
-        };
-        deleteButton.Y = (rsp.Height - deleteButton.Height) / 2;
-        deleteButton.MouseUp += (sender, e) =>
-        {
-            parentConfig.DeleteItemConfig(itemConfig);
-            // Refresh the right side to remove the deleted item
-            SelectOrganizerConfig(parentConfig, leftSideContent, button);
-        };
-        itemArea.Add(deleteButton);
-        itemArea.ForceSizeUpdate();
-        return itemArea;
-
-    }
 
     private void ShowCharacterImportContextMenu()
     {
